requirements.txt
blinker==1.9.0
cachetools==5.5.2
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
google-auth==2.38.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==1.1.0
rsa==4.9
Werkzeug==3.1.3